c_con/requirements.txt
# Python dependencies
fastapi
uvicorn
pydantic
sqlalchemy
alembic
passlib[bcrypt]
python-dotenv
